+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All Notable changes to `kevintweber/gauges` will be documented in this file.
 
 Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) principles.
 
+## [0.4] - 2017-03-16
+
+### Added
+- group parameter to `content` API call. (Thanks to David Clarke)
+- `browserstats` API call.
+
+### Updated
+- Updated dependencies to allow PHP7.
+- Updated unit tests to use new PHPUnit class structure.
+
 ## [0.3.1] - 2016-06-18
 
 ### Updated